refactor file layout and fix redirecting glitches
