+++ app/_layout.tsx
@@ -16,6 +16,8 @@ const InitialLayout = () => {
     // Todo: Remove when adding back authentication
     let devSession = true;
 
+    console.log(session);
+
     // Check if the path is in the authenticated group
     const isAuthGroup = segments[0] === "(authenticated)";
 
@@ -23,13 +25,13 @@ const InitialLayout = () => {
     // && userStatus === "Returning" // maybe this segment is not needed as long as we have a way to decipher a new user
     if (session && !isAuthGroup) {
       // router.replace("/(auth)/onboarding");
-      router.replace("/(authenticated)");
+      router.replace("/(authenticated)/(tabs)/home");
       // return;
     } else if (session && userStatus === "New") {
       // router.replace("/(auth)/language-selection");
       // return;
     } else if (!session) {
-      router.replace("/(auth)");
+      router.navigate("/");
       // router.replace("/(authenticated)");
     }
   }, [session, initialized, userStatus]);
