+++ app/index.tsx
@@ -1,8 +1,45 @@
+import { View, Text, Image } from "react-native";
 import React from "react";
-import { Redirect } from "expo-router";
+import { tw } from "@/utils/tailwind";
+import { useRouter } from "expo-router";
+import ActionButton from "@/components/ActionButton";
+import { StatusBar } from "expo-status-bar";
 
 const Index = () => {
-  return <Redirect href="/(auth)/" />;
+  const router = useRouter();
+
+  return (
+    <View
+      style={tw.style(
+        "relative flex-1 items-center justify-center bg-primary",
+      )}>
+      <StatusBar style="light" />
+      <View style={tw.style("flex items-center justify-center gap-2")}>
+        <Image
+          source={require("@/assets/images/icon.png")}
+          resizeMode="contain"
+          style={tw.style("h-52 w-64")}
+        />
+        <Text
+          style={tw.style(
+            "text-3xl font-semibold tracking-widest text-highlight",
+          )}>
+          Lingua
+        </Text>
+        <Text
+          style={tw.style(
+            "text-xl font-medium tracking-widest text-highlightAccent",
+          )}>
+          Speak Freely, Text Clearly
+        </Text>
+      </View>
+      <View style={tw.style("absolute bottom-10 w-[100%] items-center")}>
+        <ActionButton onPress={() => router.push("/(auth)/phone-signup")}>
+          Start Messaging
+        </ActionButton>
+      </View>
+    </View>
+  );
 };
 
 export default Index;
